fix regex

+++ deploy.sh
@@ -4,7 +4,6 @@ cd ~/game
 git pull
 
 deployFolder="deploy"
-htmlMinifierArgs="--collapse-whitespace --remove-comments --remove-optional-tags --remove-redundant-attributes --remove-script-type-attributes --remove-tag-whitespace --use-short-doctype --minify-css true --minify-js true"
 
 if [ -d "./$deployFolder" ]
 then
@@ -18,21 +17,19 @@ do
 	ln -s "$(pwd)/$name" "./$deployFolder/$name"
 done
 
-for i in $(find . -maxdepth 1 -type f -name "*.html")
-do
-    python3 ./min-names.py "$i" > "./$deployFolder/$i"
-    npx html-minifier $htmlMinifierArgs "./$deployFolder/$i" >> "./deploy/$i"
-done
-
-for i in $(find . -maxdepth 1 -type f -name "*.js" ! -name "*.min.*")
+for i in $(find . -maxdepth 1 -type f -name "*.js" -o -name "*.html" ! -name "*.min.*")
 do
 	python3 ./min-names.py "$i" > "./$deployFolder/$i"
-	npx minify "./$deployFolder/$i" > "./$deployFolder/${i%.*}.min.js"
-	rm "./$deployFolder/$i"
+	if [[ $i == *html ]] # .min.html does not work
+	then
+		npx minify "./$deployFolder/$i" >> "./$deployFolder/$i"
+	else {
+		npx minify "./$deployFolder/$i" > "./$deployFolder/${i%.*}.min.js"
+		rm "./$deployFolder/$i"
+	}
+	fi
 done
 
 printf "Running server."
 cd ./$deployFolder
 npx serve -p 6621 > /dev/null
-
-#npx regex-replace '\w*(?<!\.min\.)js' 'min.js' ./main.js --filecontents
